extract `getSchemaVersion` to shared db utils
